Add files via upload
This project was completed in 2022 as part of my B.Tech curriculum and represents my first step into the exciting world of Machine Learning. The primary goal was to explore and implement ML concepts while improving on existing CKD prediction projects by experimenting with different algorithms and techniques. It provided a hands-on opportunity to deepen my understanding of ML while tackling a real-world problem.